refactor(package): tableGroup 패키지 구조 리팩토링
tableGroup 패키지
- 기존에 order 패키지 내에 존재하였지만, 사실상 tableGroup은 혼자서 움직이기 때문에 다른 order 관련 도메인과 함께 존재해야 할 이유가 없어 리팩토링
