+++ src/main/java/kitchenpos/tablegroup/domain/TableGroup.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package kitchenpos.tablegroup.domain;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -7,6 +7,7 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import kitchenpos.order.domain.Order;
 
 @Entity
 public class TableGroup {
