+++ src/main/java/kitchenpos/tablegroup/dto/TableGroupRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.dto;
+package kitchenpos.tablegroup.dto;
 
 import java.util.List;
