+++ src/test/java/kitchenpos/order/acceptance/TableGroupAcceptanceTest.java
@@ -14,8 +14,8 @@ import io.restassured.response.Response;
 import java.util.Arrays;
 import kitchenpos.common.acceptance.AcceptanceTest;
 import kitchenpos.ordertable.dto.OrderTableResponse;
-import kitchenpos.order.dto.TableGroupRequest;
-import kitchenpos.order.dto.TableGroupResponse;
+import kitchenpos.tablegroup.dto.TableGroupRequest;
+import kitchenpos.tablegroup.dto.TableGroupResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
