+++ src/test/java/kitchenpos/order/acceptance/TableGroupRestAssured.java
@@ -3,7 +3,7 @@ package kitchenpos.order.acceptance;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import kitchenpos.order.dto.TableGroupRequest;
+import kitchenpos.tablegroup.dto.TableGroupRequest;
 import org.springframework.http.MediaType;
 
 public class TableGroupRestAssured {
